+++ src/test/resources/bot.properties
@@ -1,1 +1,1 @@
-bot.token=your bot token goes here
+bot.token=NjMzNzU0MjQyMDk1NTEzNjIw.XaYj5w.iXePCjIXkCnk667fm7lz97CsMaw
