--- test/java/events/MessageEvent.java
@@ -1,8 +1,0 @@
-package events;
-import java.util.*;
-
-interface MessageEvent
-{
-    String doAction(List<String> Parsedmsg);
-    void setParameters(List<String> Parsedmsg);
-}
