--- test/resources/LolChampRoles.properties
@@ -1,5 +1,0 @@
-mid
-top
-jungle
-adc
-support
