--- test/resources/LolGameModes.properties
@@ -1,8 +1,0 @@
-5v5
-3v3
-ARAM
-TFT
-Odyssey
-Blitz
-OFA
-ARURF
