--- test/resources/bot.properties
@@ -1,1 +1,0 @@
-bot.token=NjMzNzU0MjQyMDk1NTEzNjIw.XaYj5w.iXePCjIXkCnk667fm7lz97CsMaw
